cpukit/libdl: Address -Wsign-compare warnings
This warning occurs when comparing a signed variable to an unsigned one.
This addresses warnings that only occurred on 64-bit targets.  For the
ones which only appeared on 64-bit targets, the cause was frequently
a mismatch when comparing a combination off_t, ssize_t, and int.

+++ cpukit/libdl/rtl-archive.c
@@ -1134,7 +1134,7 @@ rtems_rtl_obj_archive_find_obj (int                     fd,
     *osize = 0;
   }
 
-  while (*ooffset < fsize)
+  while (*ooffset < (int64_t)fsize)
   {
     /*
      * Clean up any existing data.
